chore(middleware-sdk-s3): split region redirect middlewares in different files
