+++ packages/middleware-sdk-s3/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./check-content-length-header";
+export * from "./region-redirect-endpoint-middleware";
 export * from "./region-redirect-middleware";
 export * from "./s3Configuration";
 export * from "./throw-200-exceptions";
